requirements.txt
torch
torchvision
onnxruntime
pillow
numpy